requirements.txt
black
brightdata
httpx
miniagents
openai
python-dotenv
selenium
